+++ tailwind.config.ts
@@ -14,6 +14,7 @@ const config: Config = {
         serif: ["var(--font-serif)", "serif"],
         sans: ["var(--font-space)", "sans-serif"],
         vazir: ["var(--font-vazir)", "sans-serif"],
+        iransans: ["IRANSansX", "sans-serif"], // <--- Added this line
       },
       colors: {
         paper: "#F2F0E9", // رنگ کاغذ کاهی/استخوانی